+++ src/plugins/data_source/public/types.ts
@@ -4,7 +4,7 @@
  */
 
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
-export interface DataSourcePublicPluginSetup {}
+export interface DataSourcePluginSetup {}
 
 // eslint-disable-next-line @typescript-eslint/no-empty-interface
-export interface DataSourcePublicPluginStart {}
+export interface DataSourcePluginStart {}