+++ src/plugins/index_pattern_management/public/components/index_pattern_table/index_pattern_table.tsx
@@ -44,7 +44,7 @@ import { FormattedMessage } from '@osd/i18n/react';
 import { withRouter, RouteComponentProps } from 'react-router-dom';
 import React, { useState, useEffect } from 'react';
 import { i18n } from '@osd/i18n';
-import { useMount } from 'react-use';
+import { useEffectOnce, useMount } from 'react-use';
 import {
   reactRouterNavigate,
   useOpenSearchDashboards,
@@ -113,6 +113,9 @@ export const IndexPatternTable = ({ canSave, history }: Props) => {
   const [remoteClustersExist, setRemoteClustersExist] = useState<boolean>(false);
   const [isLoadingSources, setIsLoadingSources] = useState<boolean>(!dataSourceEnabled);
   const [isLoadingIndexPatterns, setIsLoadingIndexPatterns] = useState<boolean>(true);
+  const [isColumnDataLoaded, setIsColumnDataLoaded] = useState(false);
+
+  const { columns: columnRegistry } = indexPatternManagementStart;
 
   useMount(() => {
     setBreadcrumbs(getListBreadcrumbs());
@@ -154,6 +157,11 @@ export const IndexPatternTable = ({ canSave, history }: Props) => {
     );
   };
 
+  const loadColumnData = async () => {
+    await Promise.all(columnRegistry.getAll().map((column) => column.loadData()));
+    setIsColumnDataLoaded(true);
+  };
+
   useEffect(() => {
     if (!dataSourceEnabled) {
       getIndices({ http, pattern: '*', searchClient }).then((dataSources) => {
@@ -165,6 +173,10 @@ export const IndexPatternTable = ({ canSave, history }: Props) => {
       );
     }
   }, [http, creationOptions, searchClient, dataSourceEnabled]);
+
+  useEffectOnce(() => {
+    loadColumnData();
+  });
 
   chrome.docTitle.change(title);
 
@@ -198,6 +210,13 @@ export const IndexPatternTable = ({ canSave, history }: Props) => {
       dataType: 'string' as const,
       sortable: ({ sort }: { sort: string }) => sort,
     },
+    ...columnRegistry.getAll().map((column) => {
+      return {
+        ...column.euiColumn,
+        sortable: false,
+        'data-test-subj': `indexPatternTableColumn-${column.id}`,
+      };
+    }),
   ];
 
   const createButton = canSave ? (