+++ src/plugins/index_pattern_management/public/index.ts
@@ -55,4 +55,4 @@ export {
 
 export { DefaultFormatEditor } from './components/field_editor/components/field_format_editor';
 
-export { MlCardState } from './types';
+export { MlCardState, IndexPatternTableColumn, IndexPatternTableRecord } from './types';